Initial commit.
First overall setup for the component, does not do much yet.

Right now its capable of reading the gearbox pins and determining the
currently selected gear, provides rpm information.

The left hand side in the simulation UI allows to set the gearbox state
as seen from the "mahine side", i.e. that is what the real component
will see, it is controlled by the simulation comp. Use the slider
to modify the state of the simulated gearbox, the speed bar below
will show you what supported speed is currently selected.

The right hand side is the "real" component that will monitor the
gearbox pins (pin status indicated by LEDs in the UI), the component will
interpret the pins and show the current speed of the machine (speed bar
on the right hand side).

The "desired speed" selector is currently not doing anything, later it
will pass desired speed values to the real component via the
spindle_speed_in_abs pin. This will instruct the component to perform
a gearbox change by commanding the control pins, which is actually
the main purpose of this component.