Implemented overshoot/endpoint protection
This is a safety measure to prevent hardware damage. Basically, if it
ever happens that we wanted to go to the center, but either missed our
target or did chose the wrong direction, then we could end up reaching
the physical end point of the shift, but not reaching our target status
pin combination.

In this situation we know that we can't go any further, doing so could
damage the plastic gears that provide us with the pin status. So if such
a condition is ever reached, we immediately de-energize the shaft
motor and restart the shifting logic.